refactor: extract event page client component
- Move client-side logic to EventPageClient.tsx
- Keep page.tsx as server component for data fetching
- Improve separation of concerns between server and client code
